Impose maximum level restriction for BatchedTensors
We support 64 total levels. This is done so that we can represent lists
of levels as a bitset that fits into a single `int64_t` and is a
reasonable upper bound because we only support (physical) tensors of up
to 64 dimensions with vmap (see kVmapMaxTensorDims).

Test Plan:
`./build/bin/vmap_test`. One day we'll test this with the vmap Python
API.

ghstack-source-id: 260ba3a1baa1d40ba6de37f54e04ee968551f20e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39580